requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
aiomysql==0.2.0
cryptography==42.0.5
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
scikit-learn==1.4.1.post1
numpy==1.26.4